requirements.txt
torch==2.0.1
torch-geometric==2.3.1
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
fastapi==0.100.0
uvicorn==0.23.1
streamlit==1.26.0
requests==2.31.0
faiss-cpu==1.7.4
matplotlib==3.7.2
plotly==5.15.0
pyyaml
